% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{The great hello function}
\usage{
hello(name)
}
\arguments{
\item{name}{name to say hello}
}
\value{
A string
}
\description{
This funtion just prints "HEEEEllo \if{html}{\out{<Name>}}"
}
\examples{
hello("Save")
}
